Change directory webview net log files are stored in
Change directory to files/aw_net_logs so that the file provider can share them from the files directory. Also remove duplicate code to improve consistency and readability.

Also fixes a bug in startConnectNetLogService where the method only runs when sIsLogging is disabled, whereas it should run when sIsLogging is enabled.

Bug: 40653078
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5576815
Commit-Queue: Thomas Bull <thomasbull@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311228}

+++ android_webview/javatests/src/org/chromium/android_webview/test/services/AwNetLogServiceTest.java
@@ -43,9 +43,12 @@ public class AwNetLogServiceTest {
 
     @After
     public void tearDown() {
-        File[] files = AwNetLogService.getAllNetLogFiles();
-        for (File file : files) {
-            file.delete();
+        File directory = AwNetLogService.getNetLogFileDirectory();
+        File[] files = directory.listFiles();
+        if (files != null) {
+            for (File file : files) {
+                file.delete();
+            }
         }
     }
 
@@ -68,14 +71,16 @@ public class AwNetLogServiceTest {
             }
         }
 
-        Assert.assertEquals(1, AwNetLogService.getAllNetLogFiles().length);
+        File directory = AwNetLogService.getNetLogFileDirectory();
+        Assert.assertEquals(1, directory.listFiles().length);
     }
 
     @Test
     @MediumTest
     @CommandLineFlags.Add(AwSwitches.NET_LOG)
     public void testExpiredFilesDeleted() throws Throwable {
-        Assert.assertEquals(0, AwNetLogService.getAllNetLogFiles().length);
+        File directory = AwNetLogService.getNetLogFileDirectory();
+        Assert.assertEquals(0, directory.listFiles().length);
         final long expiredTime = 100000L;
         Intent intent = new Intent(ContextUtils.getApplicationContext(), AwNetLogService.class);
         try (ServiceConnectionHelper helper =
@@ -90,7 +95,7 @@ public class AwNetLogServiceTest {
                 Log.e(TAG, e.getMessage(), e);
             }
         }
-        Assert.assertEquals(1, AwNetLogService.getAllNetLogFiles().length);
+        Assert.assertEquals(1, directory.listFiles().length);
 
         final long currentTime = System.currentTimeMillis();
         try (ServiceConnectionHelper helper =
@@ -105,7 +110,7 @@ public class AwNetLogServiceTest {
                 Log.e(TAG, e.getMessage(), e);
             }
         }
-        Assert.assertEquals(1, AwNetLogService.getAllNetLogFiles().length);
+        Assert.assertEquals(1, directory.listFiles().length);
     }
 
     @Test
@@ -127,7 +132,9 @@ public class AwNetLogServiceTest {
             }
         }
 
-        File[] files = AwNetLogService.getAllNetLogFiles();
+        File directory = AwNetLogService.getNetLogFileDirectory();
+        File[] files = directory.listFiles();
+
         Assert.assertEquals(1, files.length);
 
         String fileName = files[0].getName();
